draw buffer is now superimposed and palette applied
Previously the buffer contained two screen data and the screens were
superimposed by using SDL palette.  However the previous
implementation could not handle palette change or mode change in the
middle of the screen.  Now the buffer is superimposed and palette
applied, so the SDL palette is only set on start-up.

Now the number of CRT controller objects is decreased to one.