draw video frames at proper time and improve audio-video sync
Previously the execution time was controlled by the audio playback
only.  So two or more frames might be drawn at the same time depending
on the audio buffer size and execution speed.

Now the execution time is controlled by the video routine.  If the
video of the emulated machine is 60fps, 16ms per frame, and drawing
takes 5ms, the video routine delays remaining 11ms.  The time is
corrected periodically based on the audio playback.

Audio-video sync is improved.  Previously the audio latency was about
500ms, but now it is expected to be less than 100ms.